minor changes for rebase

+++ src/packing.c
@@ -578,7 +578,7 @@ void xnn_pack_qs8_qb4w_gemm_gio_w(
   size_t k_stride,
   size_t bl,              // block size
   const uint8_t* k,       // kernel
-  const int32_t* bias,
+  const float* bias,
   const float* scale,
   void* packed_weights,
   size_t extra_bytes_bl,  // extra bytes per block
